+++ src/Image.php
@@ -135,14 +135,12 @@ class Image extends Base
     /**
      * Restore a source image.
      *
-     * @param string $hash         Hash of the image
+     * @param string $hash Hash of the image
      * @param string $organization Optional organization name
      *
      * @throws GuzzleException If the request fails for a different reason than image not found
      *
      * @return bool True if successful, false if image not found
-     *
-     * @throws \Exception
      */
     public function restoreSourceImage($hash, $organization = '')
     {
@@ -152,12 +150,46 @@ class Image extends Base
             if (404 == $e->getCode()) {
                 return false;
             }
-
             throw $e;
         }
 
         return '200' == $response->getStatusCode();
     }
+
+
+    /**
+     * Copy a source image to another org
+     *
+     * @param string $hash Hash of the image
+     * @param string $destinationOrg The destination organization
+     * @param bool $overwrite If an existing image should be overwritten
+     * @param string $organization Optional organization name
+     *
+     * @throws GuzzleException If the request fails for a different reason than image not found
+     *
+     * @return bool True if successful, false if source image not found
+     */
+    public function copySourceImage($hash, $destinationOrg, $overwrite = true, $organization = '')
+    {
+        try {
+            $headers = ['Destination' => $destinationOrg];
+            if ($overwrite === false) {
+                $headers['Overwrite'] = 'F';
+            }
+            $response = $this->call('COPY',
+                implode('/', [self::SOURCEIMAGE_RESOURCE, $this->getOrganization($organization), $hash]),
+                ['headers' => $headers]
+            );
+        } catch (GuzzleException $e) {
+            if (404 == $e->getCode()) {
+                return false;
+            }
+            throw $e;
+        }
+        $statusCode = $response->getStatusCode();
+        return  $statusCode >= 200 && $statusCode < 300;
+    }
+
     /**
      * Delete source images by binaryhash.
      *